Reparados todos los "0.0"

+++ lawen-ejb/src/main/java/entities/cardiovascular/PreingresoD1.java
@@ -43,28 +43,28 @@ public class PreingresoD1 implements Serializable {
     @ManyToOne(cascade = ALL, fetch = FetchType.LAZY)
     private TipoPrograma tipoprograma;
     private String referido;//profesional referido
-    private int pass1 ;// pa sistolica sentado 
-    private int pass2;// pa sistolica sentado 
-    private int pads1;// pa diastolica sentado
-    private int pads2;// pa diastolica sentado
-    private int pasp1;// pa sistolica pie
-    private int pasp2;// pa sistolica pie 
-    private int padp1;// pa diastolica pie 
-    private int padp2;// pa diastolica pie
+    private Integer pass1 ;// pa sistolica sentado 
+    private Integer pass2;// pa sistolica sentado 
+    private Integer pads1;// pa diastolica sentado
+    private Integer pads2;// pa diastolica sentado
+    private Integer pasp1;// pa sistolica pie
+    private Integer pasp2;// pa sistolica pie 
+    private Integer padp1;// pa diastolica pie 
+    private Integer padp2;// pa diastolica pie
     private String antecedentespersocvc;
     private String antecedentesfamicvc;
-    private double peso;
-    private double talla;
+    private Float peso;
+    private Float talla;
     private String trataactual;//tratamiento farmacologico
     private boolean citarmed; //se cita al medico para el siguiente
     private boolean citar6;//se cita a 6 meses al paciente
     private boolean citard2;//cita a d2
-    private int pulso;
+    private Integer pulso;
     private String observaciones;
-    private int paref1parado;//presion arterial referencia parado
-    private int paref2parado;
-    private int paref1sentado;//presion arterial referencia sentado
-    private int paref2sentado;
+    private Integer paref1parado;//presion arterial referencia parado
+    private Integer paref2parado;
+    private Integer paref1sentado;//presion arterial referencia sentado
+    private Integer paref2sentado;
     private boolean alta;//dado de alta
     private boolean derivacionicv;//deriva a icv
     @Temporal(javax.persistence.TemporalType.DATE)
@@ -113,83 +113,83 @@ public class PreingresoD1 implements Serializable {
         this.referido = referido;
     }
 
-    public int getPass1() {
+    public Integer getPass1() {
         return pass1;
     }
 
-    public void setPass1(int pass1) {
+    public void setPass1(Integer pass1) {
         this.pass1 = pass1;
     }
 
-    public int getPass2() {
+    public Integer getPass2() {
         return pass2;
     }
 
-    public void setPass2(int pass2) {
+    public void setPass2(Integer pass2) {
         this.pass2 = pass2;
     }
 
-    public int getPads1() {
+    public Integer getPads1() {
         return pads1;
     }
 
-    public void setPads1(int pads1) {
+    public void setPads1(Integer pads1) {
         this.pads1 = pads1;
     }
 
-    public int getPads2() {
+    public Integer getPads2() {
         return pads2;
     }
 
-    public void setPads2(int pads2) {
+    public void setPads2(Integer pads2) {
         this.pads2 = pads2;
     }
 
-    public int getPasp1() {
+    public Integer getPasp1() {
         return pasp1;
     }
 
-    public void setPasp1(int pasp1) {
+    public void setPasp1(Integer pasp1) {
         this.pasp1 = pasp1;
     }
 
-    public int getPasp2() {
+    public Integer getPasp2() {
         return pasp2;
     }
 
-    public void setPasp2(int pasp2) {
+    public void setPasp2(Integer pasp2) {
         this.pasp2 = pasp2;
     }
 
-    public int getPadp1() {
+    public Integer getPadp1() {
         return padp1;
     }
 
-    public void setPadp1(int padp1) {
+    public void setPadp1(Integer padp1) {
         this.padp1 = padp1;
     }
 
-    public int getPadp2() {
+    public Integer getPadp2() {
         return padp2;
     }
 
-    public void setPadp2(int padp2) {
+    public void setPadp2(Integer padp2) {
         this.padp2 = padp2;
     }
 
-    public double getPeso() {
+    public Float getPeso() {
         return peso;
     }
 
-    public void setPeso(double peso) {
+    public void setPeso(Float peso) {
         this.peso = peso;
     }
 
-    public double getTalla() {
+    public Float getTalla() {
         return talla;
     }
 
-    public void setTalla(double talla) {
+    public void setTalla(Float talla) {
         this.talla = talla;
     }
 
@@ -225,11 +225,11 @@ public class PreingresoD1 implements Serializable {
         this.citard2 = citard2;
     }
 
-    public int getPulso() {
+    public Integer getPulso() {
         return pulso;
     }
 
-    public void setPulso(int pulso) {
+    public void setPulso(Integer pulso) {
         this.pulso = pulso;
     }
 
@@ -241,35 +241,35 @@ public class PreingresoD1 implements Serializable {
         this.observaciones = observaciones;
     }
 
-    public int getParef1parado() {
+    public Integer getParef1parado() {
         return paref1parado;
     }
 
-    public void setParef1parado(int paref1parado) {
+    public void setParef1parado(Integer paref1parado) {
         this.paref1parado = paref1parado;
     }
 
-    public int getParef2parado() {
+    public Integer getParef2parado() {
         return paref2parado;
     }
 
-    public void setParef2parado(int paref2parado) {
+    public void setParef2parado(Integer paref2parado) {
         this.paref2parado = paref2parado;
     }
 
-    public int getParef1sentado() {
+    public Integer getParef1sentado() {
         return paref1sentado;
     }
 
-    public void setParef1sentado(int paref1sentado) {
+    public void setParef1sentado(Integer paref1sentado) {
         this.paref1sentado = paref1sentado;
     }
 
-    public int getParef2sentado() {
+    public Integer getParef2sentado() {
         return paref2sentado;
     }
 
-    public void setParef2sentado(int paref2sentado) {
+    public void setParef2sentado(Integer paref2sentado) {
         this.paref2sentado = paref2sentado;
     }
